move download-zip to its own folder
