+++ lib/core/dialogs/download-zip/download-zip.dialog.ts
@@ -18,8 +18,8 @@
 import { Component, Inject, OnInit, ViewEncapsulation } from '@angular/core';
 import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material';
 import { DownloadEntry, NodeEntry } from '@alfresco/js-api';
-import { LogService } from '../services/log.service';
-import { DownloadZipService } from '../services/download-zip.service';
+import { LogService } from '../../services/log.service';
+import { DownloadZipService } from '../../services/download-zip.service';
 
 @Component({
     selector: 'adf-download-zip-dialog',
