+++ lib/core/directives/node-download.directive.ts
@@ -18,7 +18,7 @@
 import { Directive, Input, HostListener } from '@angular/core';
 import { MatDialog } from '@angular/material';
 import { AlfrescoApiService } from '../services/alfresco-api.service';
-import { DownloadZipDialogComponent } from '../dialogs/download-zip.dialog';
+import { DownloadZipDialogComponent } from '../dialogs/download-zip/download-zip.dialog';
 import { NodeEntry } from '@alfresco/js-api';
 
 /**
